use Metadata() and dedicated column instead of AddMetadata()
