+++ app/process/bikelanes/bikelanes.lua
@@ -9,8 +9,8 @@ require("ToNumber")
 -- require("PrintTable")
 require("AddAddress")
 require("MergeArray")
-require("AddMetadata")
-require("AddUrl")
+require("Metadata")
+
 require("HighwayClasses")
 require("AddSkipInfoToHighways")
 require("AddSkipInfoByWidth")
@@ -22,6 +22,7 @@ local table = osm2pgsql.define_table({
   ids = { type = 'any', id_column = 'osm_id', type_column = 'osm_type' },
   columns = {
     { column = 'tags', type = 'jsonb' },
+    { column = 'meta', type = 'jsonb'},
     { column = 'geom', type = 'linestring' },
   }
 })
@@ -31,6 +32,7 @@ local skipTable = osm2pgsql.define_table({
   ids = { type = 'any', id_column = 'osm_id', type_column = 'osm_type' },
   columns = {
     { column = 'tags', type = 'jsonb' },
+    { column = 'meta', type = 'jsonb'},
     { column = 'geom', type = 'linestring' },
   }
 })
@@ -247,12 +249,11 @@ function osm2pgsql.process_way(object)
     "traffic_sign",
   })
   FilterTags(object.tags, allowed_tags)
-  AddMetadata(object)
-  AddUrl("way", object)
 
   if object.tags._skip then
     skipTable:insert({
       tags = object.tags,
+      meta = Metadata(object),
       geom = object:as_linestring()
     })
   else
@@ -261,6 +262,7 @@ function osm2pgsql.process_way(object)
     object.tags._skipNotes = nil
     table:insert({
       tags = object.tags,
+      meta = Metadata(object),
       geom = object:as_linestring()
     })
   end
