+++ app/process/bikelanes/bikelanesNew.lua
@@ -5,8 +5,8 @@ require("ToNumber")
 -- require("PrintTable")
 require("AddAddress")
 require("MergeArray")
-require("AddMetadata")
-require("AddUrl")
+require("Metadata")
+
 require("HighwayClasses")
 require("RoadWidth")
 require("AddSkipInfoToHighways")
