+++ app/process/lit/lit.lua
@@ -5,8 +5,8 @@ require("ToNumber")
 -- require("PrintTable")
 require("AddAddress")
 require("MergeArray")
-require("AddMetadata")
-require("AddUrl")
+require("Metadata")
+
 require("HighwayClasses")
 require("AddSkipInfoToHighways")
 require("AddSkipInfoByWidth")
@@ -17,6 +17,7 @@ local table = osm2pgsql.define_table({
   ids = { type = 'any', id_column = 'osm_id', type_column = 'osm_type' },
   columns = {
     { column = 'tags', type = 'jsonb' },
+    { column = 'meta', type = 'jsonb'},
     { column = 'geom', type = 'linestring' },
   }
 })
@@ -138,13 +139,11 @@ function osm2pgsql.process_way(object)
     "cycleway:width", -- experimental
   })
   FilterTags(object.tags, allowed_tags)
-  AddMetadata(object)
-  AddUrl("way", object)
-
   if object.tags._skip then
   else
     table:insert({
       tags = object.tags,
+      meta = Metadata(object),
       geom = object:as_linestring()
     })
   end
