+++ app/process/places/places_todoList.lua
@@ -5,14 +5,15 @@ require("ToNumber")
 -- require("PrintTable")
 require("AddAddress")
 require("MergeArray")
-require("AddMetadata")
-require("AddUrl")
+require("Metadata")
+
 
 local table = osm2pgsql.define_table({
   name = 'places_todoList',
   ids = { type = 'any', id_column = 'osm_id', type_column = 'osm_type' },
   columns = {
     { column = 'tags', type = 'jsonb' },
+    { column = 'meta', type = 'jsonb'},
     { column = 'geom', type = 'point' },
   }
 })
@@ -55,17 +56,16 @@ local function ProcessTags(object)
     "population:date", "admin_level" })
   FilterTags(object.tags, allowed_tags)
   -- ToNumber(object.tags, Set({ "population" }))
-  AddMetadata(object)
 end
 
 function osm2pgsql.process_node(object)
   if not ContinueProcess(object) then return end
 
   ProcessTags(object)
-  AddUrl("node", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_point()
   })
 end
@@ -75,10 +75,10 @@ function osm2pgsql.process_way(object)
   if not object.is_closed then return end
 
   ProcessTags(object)
-  AddUrl("way", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_polygon():centroid()
   })
 end
@@ -88,10 +88,10 @@ function osm2pgsql.process_relation(object)
   if not object.tags.type == 'multipolygon' then return end
 
   ProcessTags(object)
-  AddUrl("relation", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_multipolygon():centroid()
   })
 end
