+++ app/process/poiClassification.lua
@@ -6,8 +6,8 @@ require("FilterTags")
 -- require("PrintTable")
 require("AddAddress")
 require("MergeArray")
-require("AddMetadata")
-require("AddUrl")
+require("Metadata")
+
 -- Shared:
 require("ShoppingAllowedListWithCategories")
 
@@ -17,6 +17,7 @@ local table = osm2pgsql.define_table({
   ids = { type = 'any', id_column = 'osm_id', type_column = 'osm_type' },
   columns = {
     { column = 'tags', type = 'jsonb' },
+    { column = 'meta', type = 'jsonb'},
     { column = 'geom', type = 'point' },
   }
 })
@@ -59,17 +60,16 @@ local function ProcessTags(object)
   local allowed_addr_tags = AddAddress(object.tags)
   local allowed_tags = Set(MergeArray({ "name", "category", "type" }, allowed_addr_tags))
   FilterTags(object.tags, allowed_tags)
-  AddMetadata(object)
 end
 
 function osm2pgsql.process_node(object)
   if ExitProcessing(object) then return end
 
   ProcessTags(object)
-  AddUrl("node", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_point()
   })
 end
@@ -79,10 +79,10 @@ function osm2pgsql.process_way(object)
   if not object.is_closed then return end
 
   ProcessTags(object)
-  AddUrl("way", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_polygon():centroid()
   })
 end
@@ -92,10 +92,10 @@ function osm2pgsql.process_relation(object)
   if not object.tags.type == 'multipolygon' then return end
 
   ProcessTags(object)
-  AddUrl("relation", object)
 
   table:insert({
     tags = object.tags,
+    meta = Metadata(object),
     geom = object:as_multipolygon():centroid()
   })
 end
